lib/pycore: fix the Python log() function work when no arguments are passed
proxyLog() called log() with args=None as the last argument when only a
format string was provided in the proxyLog() call. Unfortunately this does
did not work as the underlying z_py_log() C function was under the
assumption that the args argument won't be present if there are no args.

This patch also adds a test checking that no exceptions are generated when
log() is called with None as the last argument.