Added initial LuaObject support.
To be able to use wrapped userdata as a normal table in Lua.
Every userdata has a variable table which is stored in the registry, this table's
metatable is populated with the C functions. The userdata's metatable
holds only the __index, __newindex methods which are proxies between
the userdata and the userdata's variable table. I should write the __gc
method, too. I have to create the bind method, which binds an already existing Lua table
as a userdata's variable table.
Next step is to create semi-automatic bindings between lua variables and
userdata wrapped C object variables.